refactor(mcp): delegate tool handlers to core operations module
Move MCP business logic to src/transports/mcp/handlers.ts which calls
remember/recall/forget from src/core/operations.ts. Keep src/mcp/tools.ts
as a backward-compat re-export shim so existing imports and .mcp.json
entry point are unaffected.
